versions/v1.0 - Initial Release/requirements.txt
flask==3.0.3
pandas==2.2.2
openpyxl==3.1.2
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
watchdog==4.0.0
requests==2.31.0
python-dotenv==1.0.1
reportlab==4.2.0
xlrd==2.0.1
